+++ frontend/tests/components/scan/ScanForm/requests/ScanForm.url.test.jsx
@@ -26,9 +26,6 @@ jest.mock("../../../../../src/components/scan/utils/RecentScans", () =>
 );
 
 describe("test ScanForm component with URLs", () => {
-  /* those tests could require lots of time, jest.setTimeout doesn't work on async function.
-        use the second param of async instead.
-    */
 
   beforeAll(() => {
     axios.post.mockImplementation(() =>
